requirements.txt
Alfred-Workflow==1.36
certifi==2018.10.15
chardet==3.0.4
idna==2.7
requests==2.20.0
ujson==1.35
urllib3==1.24.1
